Enable more errorprone checks (#85)
Enable more checks from the errorprone static analyzer, and fix some of
the bugs/recommendations that it detected.

+++ src/main/java/com/team766/simulator/ui/Metrics.java
@@ -57,7 +57,7 @@ public class Metrics extends JPanel {
                 System.out.println(
                         String.format(
                                 "(%s, %f): %f",
-                                source.getName(), selectedTime, source.get(1, index)));
+                                source.getName(), selectedTime, (Double) source.get(1, index)));
             }
         }
